[#30]fix: 충돌 해결, member -> account 변경에 따른 경로 수정
